# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_source_experts__slug__sources_upload_post": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_source_experts__slug__sources_upload_post",
        "type": "object"
      },
...
      "ExpertAccess": {
        "description": "The caller's relationship to an expert they can read (migration 031).\n\nEchoed on every expert response so a client can hide the controls a viewer\ncannot use. It is a rendering hint only \u2014 every mutating route re-checks\nownership itself.",
        "enum": [
          "owner",
          "viewer"
        ],
        "title": "ExpertAccess",
        "type": "string"
      },
      "ExpertAvatar": {
        "description": "A rendering recipe, not an image. See peritus.experts.avatar for why.\n\n``seed`` is optional; the client derives it when absent. ``hue`` is kept for\nolder clients and is always null \u2014 there are no per-expert colours, and a\nvalue sent here is discarded rather than refused.",
        "properties": {
          "hue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hue"
          },
          "seed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "style": {
            "title": "Style",
            "type": "string"
          }
        },
        "required": [
          "style"
        ],
        "title": "ExpertAvatar",
        "type": "object"
      },
      "ExpertPictureOut": {
        "description": "A found, licensed picture of the expert's *subject* (migration 027).\n\nNot the avatar and not a headshot: the avatar is the owner's rendering\nrecipe and stays authoritative, while this is what a fresh expert arrives\nwith. The bytes are at ``GET /experts/{slug}/picture``; ``version`` is the\nshort content hash that goes in its ``?v=`` so an immutable cache is safe.\n\nThe provenance fields are not decoration. CC BY and CC BY-SA require\nattribution wherever the picture is shown as an identity, so a client that\nrenders this has to render the credit too \u2014 ``attribution_required`` says\nwhen, and ``file_page_url`` is where the licence is actually stated.",
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "attribution_required": {
            "default": true,
            "title": "Attribution Required",
            "type": "boolean"
          },
          "file_page_url": {
            "title": "File Page Url",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "license_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Url"
          },
          "page_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Url"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "width",
          "height",
          "provider",
          "license",
          "file_page_url"
        ],
        "title": "ExpertPictureOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ShareAcceptOut": {
        "description": "Where to send someone who has just opened a link while signed in.",
        "properties": {
          "access": {
            "$ref": "#/components/schemas/ExpertAccess"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "access"
        ],
        "title": "ShareAcceptOut",
        "type": "object"
      },
      "ShareStateOut": {
        "description": "The owner's view of an expert's share link.\n\n``token`` is null while sharing is off. The API does not know the web app's\norigin, so the client builds the URL (``/share/{token}``).",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "uploaded_source_count": {
            "default": 0,
            "title": "Uploaded Source Count",
            "type": "integer"
          },
          "viewer_count": {
            "default": 0,
            "title": "Viewer Count",
            "type": "integer"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "ShareStateOut",
        "type": "object"
      },
      "SharedExpertOut": {
        "description": "What anyone holding a live share link sees, signed in or not.\n\nThe link preview and the share page render from this. It is the catalog\ncard's shape minus curation, plus the avatar recipe: no slug, no owner, no\nerror, no build internals \u2014 nothing the link's holder could not see after\nopening it, and nothing about who shared it.",
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertAvatar"
              },
              {
                "type": "null"
              }
            ]
          },
          "avg_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Quality"
          },
          "build_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Active"
          },
          "chunk_count": {
            "default": 0,
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "graph_expanded": {
            "default": false,
            "title": "Graph Expanded",
            "type": "boolean"
          },
          "key_concepts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Key Concepts",
            "type": "array"
          },
          "node_count": {
            "default": 0,
            "title": "Node Count",
            "type": "integer"
          },
          "persona_bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Bio"
          },
          "persona_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Name"
          },
          "picture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExpertPictureOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "readiness": {
            "default": "pending",
            "title": "Readiness",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "source_type_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Source Type Counts",
            "type": "object"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "tier",
          "created_at"
        ],
        "title": "SharedExpertOut",
        "type": "object"
      },
...
      "UploadAcceptedOut": {
        "description": "Returned the moment a payload is durably stored and queued.\n\n``job_id`` is what the client tails on ``/experts/{slug}/build/events`` \u2014\ningest progress rides the same event stream as a build.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "upload_id": {
            "title": "Upload Id",
            "type": "integer"
          }
        },
        "required": [
          "upload_id",
          "job_id",
          "title",
          "kind"
        ],
        "title": "UploadAcceptedOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Peritus API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/experts/{slug}/access": {
      "delete": {
        "description": "Remove a shared expert from the caller's workspace.\n\nThe owner cannot \"leave\" their own expert \u2014 that would be a delete, which\nis a different, destructive action with its own route.",
        "operationId": "leave_shared_expert_experts__slug__access_delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Leave Shared Expert",
        "tags": [
          "sharing"
        ]
      }
    },
...
    "/experts/{slug}/share": {
      "delete": {
        "description": "Turn the link off. Everyone who opened it loses access; their own chats\nare kept but can no longer be continued.",
        "operationId": "disable_share_experts__slug__share_delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Share",
        "tags": [
          "sharing"
        ]
      },
      "get": {
        "operationId": "get_share_experts__slug__share_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Share",
        "tags": [
          "sharing"
        ]
      },
      "put": {
        "description": "Turn the link on. Idempotent: a live link is returned, never replaced.",
        "operationId": "enable_share_experts__slug__share_put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enable Share",
        "tags": [
          "sharing"
        ]
      }
    },
    "/experts/{slug}/share/reset": {
      "post": {
        "description": "Replace the link. Everyone who opened the old one loses access.",
        "operationId": "reset_share_experts__slug__share_reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Share",
        "tags": [
          "sharing"
        ]
      }
    },
...
    "/experts/{slug}/sources/upload": {
      "post": {
        "description": "Accept a PDF or text/markdown file and queue it for ingestion.\n\n202, not 201: the source does not exist yet. What exists is a durable\npayload and a queued job, and the response says which job to watch.",
        "operationId": "upload_source_experts__slug__sources_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_experts__slug__sources_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAcceptedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Source",
        "tags": [
          "sources"
        ]
      }
    },
...
    "/share/{token}": {
      "get": {
        "description": "The share card. Readable without a session, so a link can unfurl.",
        "operationId": "get_shared_expert_share__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedExpertOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Shared Expert",
        "tags": [
          "sharing"
        ]
      }
    },
    "/share/{token}/accept": {
      "post": {
        "description": "Open the link as a signed-in user: record a grant, return the slug.\n\nIdempotent. The owner opening their own link gets no grant \u2014 they already\nhave more than one would give them.",
        "operationId": "accept_share_share__token__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareAcceptOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Share",
        "tags": [
          "sharing"
        ]
      }
    },
    "/share/{token}/picture": {
      "get": {
        "description": "The picture's bytes for the share page and its link preview.\n\nRevalidated on every use (``no-cache`` with an ETag) rather than cached\nimmutably like the signed-in picture: a 304 costs nothing, and a link that\nhas been turned off must stop serving its image too.",
        "operationId": "get_shared_picture_share__token__picture_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Shared Picture",
        "tags": [
          "sharing"
        ]
      }
    }
  }
}